fix(demo): properly renamed `ComponentWrapper` file
